feat: `is_a?`, `respond_to?` and `hash` — the reflection the Ruby half was written against
`core/kernel.rb` defines `is_a?`, `kind_of?` and `respond_to?` in Ruby, and
`core/module.rb` defines `===`, `<` and `include?` the same way. Every one of
them calls `Module#ancestors` or `#method_defined?`, and neither existed — so
all six were dead, and nothing failed, because a missing method is reported
*blocked* and the specs that would have caught it were blocked on something
else first.

Found by running a handful of expressions through `spinel run` and diffing
against CRuby, which is the check a pass count cannot be.

Four primitives answer it — `Module#ancestors`, `Class#superclass`,
`Module#method_defined?`, `Object#hash` — all four class-table or raw-storage
reads. `hash` is content for `String` and `Array` and identity otherwise, which
is Ruby's default and the contract a real `Hash` will key on.

    corpus  1023 -> 1108 passed

The same pass found four wrong answers in code no spec had reached yet:

- `Hash#fetch(k, nil)` raised `KeyError` instead of answering nil — an explicit
  nil default is a default, so it counts arguments now rather than testing one.
- `1 <=> 1.5` was nil instead of -1; `Integer#<=>` compared against `Integer`
  where Ruby compares against `Numeric`.
- `BasicObject#!` asked the object's own `==`, so an object answering true to
  everything would have been falsy. It tests identity now.
- `Array#initialize` appended where Ruby replaces.

And making `respond_to?` answer turned five refusals into failures, all real:
`Array.new(a, x)` raised `ArgumentError` where Ruby raises `TypeError`,
`Array#shift` took any number of arguments, `Array#min`/`#max` ignored a
comparison block, and `Array.new(n) { break :x }` answered the array rather than
`:x` — a `Class#new` bug, where the object placed below the frame's base
survived a `break` that named the frame.

`Kernel#to_enum` is skipped with a reason: it must answer an `Enumerator`, which
needs fibers (#26, #16). It was blocked before and fails now only because
`respond_to?` can finally say no honestly.

`scripts/verify-passes.rb` re-ran all 1,108 passes on ruby 4.0.6 and all agree.

+++ core/integer.rb
@@ -80,7 +80,7 @@ class Integer
   end
 
   def <=>(other)
-    return nil unless other.is_a?(Integer)
+    return nil unless other.is_a?(Numeric)
     return -1 if self < other
     return 1 if self > other
     0
